moves chart code to component
